refactor: move MOCK_USERS to data.ts and update import path in AddMemberModal
