+++ src/app/features/portfolio-management/components/add-member-modal.tsx
@@ -2,7 +2,7 @@ import { Button } from '@/components/ui/button';
 import type { Member } from '@/types/portfolio-management';
 import { Search } from 'lucide-react';
 import { useEffect, useState } from 'react';
-import { MOCK_USERS } from '../constants';
+import { MOCK_USERS } from '../constants/data';
 
 interface AddMemberModalProps {
   isOpen: boolean;
